Adding more debug to the subscription tests and enabling it

+++ tests/integration/cleanup.ts
@@ -37,25 +37,26 @@ describe('Cleanup', async function() {
         subscriptions = await cloudAPIClient.getSubscriptions();
         console.log(subscriptions);
         expect(subscriptions.length).to.eql(0, 'Subscriptions count');
-
-        let cloudAccounts = await cloudAPIClient.getCloudAccounts();
-        for(let i = 0; i < cloudAccounts.length; i++) {
-            const cloudAccountId = cloudAccounts[i]['id'];
-            if(cloudAccountId !== 1) {
-                console.log(`=== Starting cleanup for cloud account ${cloudAccountId} ===`);
-                await cloudAPIClient.deleteCloudAccount(cloudAccountId);
-                await cloudAPIClient.waitForCloudAccountStatus(cloudAccountId, CloudAccountStatus.deleted);
-                const cloudAccount = await cloudAPIClient.getCloudAccount(cloudAccountId);
-                expect(cloudAccount['status']).to.not.eql(CloudAccountStatus.active, 'Cloud Account Status');
-                console.log(`=== Finished cleanup for cloud account ${cloudAccountId} ===`);
-            }
-            
-        }
-        cloudAccounts = await cloudAPIClient.getCloudAccounts();
-        expect(cloudAccounts.length).to.eql(1, 'Cloud accounts count');
     });
 
-    // it('Cloud account cleanup', async () => {
-        
-    // });
+    it('Cloud account cleanup', async () => {
+        const subscriptions = await cloudAPIClient.getSubscriptions();
+        if(subscriptions.length === 0) {
+            let cloudAccounts = await cloudAPIClient.getCloudAccounts();
+            for(let i = 0; i < cloudAccounts.length; i++) {
+                const cloudAccountId = cloudAccounts[i]['id'];
+                if(cloudAccountId !== 1) {
+                    console.log(`=== Starting cleanup for cloud account ${cloudAccountId} ===`);
+                    await cloudAPIClient.deleteCloudAccount(cloudAccountId);
+                    await cloudAPIClient.waitForCloudAccountStatus(cloudAccountId, CloudAccountStatus.deleted);
+                    const cloudAccount = await cloudAPIClient.getCloudAccount(cloudAccountId);
+                    expect(cloudAccount['status']).to.not.eql(CloudAccountStatus.active, 'Cloud Account Status');
+                    console.log(`=== Finished cleanup for cloud account ${cloudAccountId} ===`);
+                }
+                
+            }
+            cloudAccounts = await cloudAPIClient.getCloudAccounts();
+            expect(cloudAccounts.length).to.eql(1, 'Cloud accounts count');
+        }
+    });
 });
